chore: update about page
