+++ data/authors/default.md
@@ -10,6 +10,13 @@ github: https://github.com/Scrump31/
 
 Scottie Crump is a Staff Test Engineer with over seven years of experience as a consultant working with clients from various sectors, including automotive, retail, telecommunications, and healthcare. He received his Bachelors and Graduate Certificate from The University of North Carolina at Charlotte.
 
-He is the author of [Simplify Testing with React Testing Library: Create maintainable tests using RTL that do not break with changes](https://www.amazon.com/dp/1800564457/ref=cm_sw_em_r_mt_dp_FYYYJ07NCGKV6T1RTVYF). He holds the [Advanced Level Agile Technical Tester (CTAL-ATT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003), [Advanced Level Test Automation Engineer (CTAL-TAE)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003) , [Foundation Level Performance Tester (CTFL-PT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003), and [Foundation Level Mobile App Tester (CTFL-MAT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003) certifications via the [International Software Testing Qualifications Board](https://www.istqb.org). He is a [Amazon Web Services Certified Cloud Practitioner](https://www.credly.com/badges/74473b93-3746-474c-8998-1d3290732c44/public_url). He is also a certified [Professional Scrum Master I](https://www.scrum.org/user/872947).
+He's the author of [Simplify Testing with React Testing Library: Create maintainable tests using RTL
+that do not break with changes](https://www.amazon.com/dp/1800564457/ref=cm_sw_em_r_mt_dp_FYYYJ07NCGKV6T1RTVYF).
+He holds the [Advanced Level Agile Technical Tester (CTAL-ATT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003),
+[Advanced Level Test Automation Engineer (CTAL-TAE)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003) ,
+[Foundation Level Performance Tester (CTFL-PT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003),
+and [Foundation Level Mobile App Tester (CTFL-MAT)](https://atsqa.org/certified-testers/profile/13fa786117484e02a22bf5603c840003) certifications via the [International Software Testing Qualifications Board](https://www.istqb.org).
+He is a [Amazon Web Services Certified Cloud Practitioner](https://www.credly.com/badges/74473b93-3746-474c-8998-1d3290732c44/public_url).
+He is also a certified [Professional Scrum Master I](https://www.scrum.org/user/872947).
 
-He has also worked with numerous students looking to break into the software industry at Coding Bootcamp settings. His background in quality and software engineering and education enables him to be a crucial cog in any client's project. In addition, Scottie has helped bring industry best practices and data-driven insights to support key business objectives.
+He has also worked with many students looking to break into the software industry at Coding Bootcamp settings. His background in quality and software engineering and education enables him to be a crucial cog in any client's project. In addition, Scottie has helped bring industry best practices and data-driven insights to support key business goals.
