+++ rehype-prism-plus.d.ts
@@ -1,0 +1,1 @@
+declare module 'rehype-prism-plus'
